- For each 10,000 points we get one more life - Energizers make earn 50 points each - Adding a little cheatcode (useful for testing)... Let you find it ! ;-)
svn path=/trunk/playground/games/kapman/; revision=778248

+++ game.h
@@ -53,7 +53,7 @@ class Game : public QObject {
 		int m_lives;
 
 		/** Player's points */
-		qreal m_points;
+		int m_points;
 
 		/** Current level */
 		int m_level;
@@ -153,7 +153,7 @@ class Game : public QObject {
 		/**
 		 * Manages the points won
 		 */
-		void winPoints(qreal p_points, qreal p_x, qreal q_y);
+		void winPoints(int p_points, qreal p_x, qreal q_y);
 
 		/**
 		 * Start the next level
