+++ kapman.cpp
@@ -136,7 +136,7 @@ void Kapman::loseLife() {
 	emit(lifeLost());
 }
 
-void Kapman::winPoints(qreal p_points, qreal p_x, qreal p_y) {
+void Kapman::winPoints(int p_points, qreal p_x, qreal p_y) {
 	// Emits a signal to the game
 	emit(sWinPoints(p_points, p_x, p_y));
 }
